Removed comment node, added KDoc comments to annotations and separated expression nodes further also adding KDoc comments
