+++ src/main/kotlin/com/valgolang/runtime/datastructures/array/ArrayExecutor.kt
@@ -1,7 +1,13 @@
 package com.valgolang.runtime.datastructures.array
 
 import com.valgolang.frontend.ast.*
+import com.valgolang.frontend.datastructures.ConstructorNode
+import com.valgolang.frontend.datastructures.DataStructureInitialiserNode
+import com.valgolang.frontend.datastructures.MethodCallNode
+import com.valgolang.frontend.datastructures.array.Array2DInitialiserNode
+import com.valgolang.frontend.datastructures.array.ArrayElemNode
 import com.valgolang.frontend.datastructures.array.ArrayType
+import com.valgolang.frontend.datastructures.array.InternalArrayMethodCallNode
 import com.valgolang.frontend.datastructures.list.ListType
 import com.valgolang.linearrepresentation.EmptyMObject
 import com.valgolang.linearrepresentation.ManimInstr
